fix(schema.ts): remove 'notNull' constraint from 'tagIds' column in the 'prompts' table to allow for null values
The 'tagIds' column in the 'prompts' table was previously defined with a 'notNull' constraint, which means it cannot have null values. However, this constraint is not necessary for this column, as it should be allowed to have null values. Therefore, the 'notNull' constraint has been removed from the 'tagIds' column definition in the 'prompts' table.