feat(migrations): add migration to enable row-level security (RLS) for tag_prompt_link_table
The migration file "20230926150108_enabled_RLS_for_tag_prompt_link_table.sql" is added to enable row-level security (RLS) for the "tag_prompt_link_table" table. This policy allows all actions for users based on the "created_by" column. The policy is set to be permissive and grants access to the public role. The policy also includes a check to ensure that only the user who created the row can access it.