feat(migrations): add migration files for adding and removing the "wrappers" extension and creating the "stripe" schema
The migration files "20230928220644_added_stripe_wrapper.sql" and "20230928221224_remote_schema.sql" are added to the "supabase/migrations" directory. These files contain SQL statements to create the "wrappers" extension with the "public" schema and the "extensions" schema respectively.

The migration file "20230928221333_remote_schema.sql" is added to the "supabase/migrations" directory. This file contains SQL statements to create the "wrappers" extension with the "extensions" schema.

The migration file "20230928221813_remote_schema.sql" is added to the "supabase/migrations" directory. This file contains SQL statements to create the "stripe" schema.