feat(migrations): add remote schema migration files of updated trigger and function
The migration files `20230929104200_remote_schema.sql` and `20230929104427_remote_schema.sql` have been added to the `supabase/migrations` directory.

`20230929104200_remote_schema.sql` contains the SQL code of updated function `create_profile_on_signup()`. This function is triggered after a new user is signed up and performs the following actions:
- Inserts a new row into the `stripe.customers` table with the `profile_id` in the `attrs` field.
- Retrieves the `id` of the newly inserted customer from the `stripe.customers` table.
- Inserts the new user's ID, email, full name, avatar URL, and the retrieved `stripe_customer_id` into the `public.profiles_table`.
- Returns the new record.

`20230929104427_remote_schema.sql` contains the SQL code of updated  trigger `on_auth_user_created` that is executed after a new row is inserted into the `auth.users` table. The trigger executes the `create_profile_on_signup()` function.

These migration files are added to implement the functionality of creating a profile for a user upon signup and linking it to a customer in the Stripe service.