feat(migrations): add subscription_status column to profiles_table
Add a new migration file to add the subscription_status column to the profiles_table. The column is of type enum and has the following possible values: 'trialing', 'active', 'paused', 'ended', 'canceled', 'past_due', 'unpaid', 'incomplete', 'incomplete_expired'. The column is set to not null and has a default value of 'trialing'.